refactor(cli): migrate erddap tooling onto discovery
Move erddap's refresher/writer/coverage/validator/emitter to earthlens.erddap.cli,
wiring the first coverage role through discovery (refresh's _COVERAGE projects
dispatch_table). Add COVERAGE_BUCKETS to the public toolkit. Relocate erddap's
CLI tests to earthlens-ocean. Shrink the guard allow-list (erddap).
