+++ libs/core/tests/cli/test_validate.py
@@ -13,7 +13,6 @@ from earthlens.cli.validate import (
     _live_ecmwf,
     _validate_bathymetry,
     _validate_drought,
-    _validate_erddap,
     _validate_goes,
     _validate_nrel,
     _validate_nwp,
@@ -306,49 +305,6 @@ class TestValidateNwp:
         _checked, issues = _validate_nwp(catalog)
         assert any("empty band map" in i for i in issues), "empty bands flagged"
         assert any("out of range" in i for i in issues), "bad cycle flagged"
-
-
-class TestValidateErddap:
-    """The ERDDAP offline lint flags the cross-row problems the model can't."""
-
-    @staticmethod
-    def _catalog(**row_overrides):
-        """A one-row fake catalog whose row carries the given overrides."""
-        from earthlens.erddap.catalog import Dataset
-
-        fields = dict(
-            server_url="https://example.org/erddap",
-            dataset_id="d",
-            protocol="tabledap",
-            variables=["a"],
-        )
-        fields.update(row_overrides)
-        return SimpleNamespace(datasets={"d": Dataset(**fields)})
-
-    def test_clean_row_has_no_issues(self):
-        """A well-formed row lints clean and is counted."""
-        checked, issues = _validate_erddap(self._catalog())
-        assert checked == 1, f"expected 1 row checked, got {checked}"
-        assert issues == [], f"clean row should have no issues, got {issues}"
-
-    def test_non_http_server_url_flagged(self):
-        """A server_url that is not http(s) is flagged."""
-        _, issues = _validate_erddap(self._catalog(server_url="ftp://x/erddap"))
-        assert any("http(s)" in i for i in issues), f"server_url not flagged: {issues}"
-
-    def test_empty_griddap_dim_names_flagged(self):
-        """A griddap row with empty dim_names is flagged."""
-        _, issues = _validate_erddap(self._catalog(protocol="griddap", dim_names=[]))
-        assert any("dim_names" in i for i in issues), f"dim_names not flagged: {issues}"
-
-    def test_flux_variable_not_in_variables_flagged(self):
-        """A flux_variables entry absent from the row's variables is flagged."""
-        _, issues = _validate_erddap(
-            self._catalog(protocol="griddap", variables=["a"], flux_variables=["b"])
-        )
-        assert any("flux_variables" in i for i in issues), (
-            f"flux typo not flagged: {issues}"
-        )
 
 
 class TestStructuralLints:
