+++ libs/core/tests/test_distribution_boundaries.py
@@ -1277,7 +1277,6 @@ class TestCliIsBackendAgnostic:
             "drought",
             "earthdata",
             "ecmwf",
-            "erddap",
             "eumetsat",
             "fabdem",
             "gbif",
